feat: Move driver function from test to source
The helper function `minimize` used in the test problems is essentially
a driver function of the MMA implementation. At some point this will be
the main minimizer routine, either as function or embedded in some class
definition, but this should reside within the source and should not be
the responsibility of the user to define.