feat: Move approximating functions into helper function
The calculation of the approximating functions through components `P` and
`Q` are shifted into a separate helper function. The operations can be
reused for the objective and constraint functions too by transposing the
derivative information when needed.

Additionally, this moves from `diags` to `diags_array` to match with the
latest scipy interface and adopts `squeeze` over `flatten` as this
(might) return a view of the underlying data rather than an explicit
copy.