feat: Move xold into TargetFunction to make `mmasub` free
The `mmasub` function can be defined without a wrapper class to track
the previous two iterates when we shift this state into the
`TargetFunction` definition. This might make more sense as this class is
storing the current `x` iterate as well, although it does hide this
history dependence from `mmasub`.

However, this implementation does explicitly invoke the `store` method,
which should be clear enough that the history is maintained elsewhere?