feat: Collect test data through callback function
Add support for user-defined callback functions to be passed into the
main MMA entry point that will be evaluated every iteration with an
(intermediate) optimization result `OptimizationResult` instance.

This functionality is leveraged to extract the iteration data used for
the test cases and allows to remove the test specific code from the main
function.